+++ support/iam/src/console_system/api/iam_cs_tenant_api.rs
@@ -80,8 +80,8 @@ impl IamCsTenantApi {
         name: Query<Option<String>>,
         desc_by_create: Query<Option<bool>>,
         desc_by_update: Query<Option<bool>>,
-        page_number: Query<u64>,
-        page_size: Query<u64>,
+        page_number: Query<u32>,
+        page_size: Query<u32>,
         ctx: TardisContextExtractor,
     ) -> TardisApiResult<TardisPage<IamTenantSummaryResp>> {
         let funs = iam_constants::get_tardis_inst();
